[assistant+shared] Extract reusable outbound proxy core into packages/egress-proxy
